update more examples

+++ mainGndMapRdr/main.cpp
@@ -19,24 +19,24 @@
 #include <cstdlib>
 
 // background rate
-const int bgRate {20};
-TestStation* testStation {};
+const int bgRate{20};
+TestStation* testStation{};
 
 // updateDataCB() -- Station's background tasks
 void updateDataCB(int)
 {
-   const double dt0 {1.0 / static_cast<double>(bgRate)};
-   const int millis {static_cast<int>(dt0 * 1000)};
+   const double dt0{1.0 / static_cast<double>(bgRate)};
+   const int millis{static_cast<int>(dt0 * 1000)};
    glutTimerFunc(millis, updateDataCB, 1);
 
    // current time
-   const double time {mixr::base::getComputerTime()};
+   const double time{mixr::base::getComputerTime()};
 
    // N-1 Time
-   static double time0 {time};
+   static double time0{time};
 
    // compute delta time
-   const double dt {static_cast<double>(time - time0)};
+   const double dt{time - time0};
    time0 = time;
 
    mixr::base::Timer::updateTimers(dt);
@@ -48,8 +48,8 @@ void updateDataCB(int)
 TestStation* builder(const std::string& filename)
 {
    // read configuration file
-   int num_errors {};
-   mixr::base::Object* obj {mixr::base::edl_parser(filename, factory, &num_errors)};
+   int num_errors{};
+   mixr::base::Object* obj{mixr::base::edl_parser(filename, factory, &num_errors)};
    if (num_errors > 0) {
       std::cerr << "File: " << filename << ", number of errors: " << num_errors << std::endl;
       std::exit(EXIT_FAILURE);
@@ -96,13 +96,13 @@ int main(int argc, char* argv[])
 
    testStation->reset();
 
-   const double end {mixr::base::getComputerTime()};
-   const double dtime {end - start};
+   const double end{mixr::base::getComputerTime()};
+   const double dtime{end - start};
    std::cout << "finished loading files: time(s) = " << dtime << std::endl;
 
    // set timer for background tasks
-   const double dt {1.0 / static_cast<double>(bgRate)};
-   const int millis {static_cast<int>(dt * 1000)};
+   const double dt{1.0 / static_cast<double>(bgRate)};
+   const int millis{static_cast<int>(dt * 1000)};
    glutTimerFunc(millis, updateDataCB, 1);
 
    testStation->createTimeCriticalProcess();
